Update updated_app.mlapp
Now that we figured out the code to accurately identify if letters should be red, green, or yellow (and account for repeated letters), I adjusted the format of the MATLAB GUI to be 5 columns when we were previously working with 4. Along with this, I altered the existing code to go through 5 columns instead of 4 like we had been working with.