Update updated_app.mlapp
updated the code to use a random list of 3000 5-letter words and pick one of those words for the user to guess thus making the game a lot more difficult. All three of us worked on this in class today, I just uploaded it. Mara primarily worked on imported the txt file while paris and I helped to brainstorm ideas about how to create each word into a string vector